+++ include/hook_fops.h
@@ -1,3 +1,3 @@
-#include "detectiontool.h"
+#include "main.h"
 
 int scan_fops(void);
