+++ include/hook_syscalls.h
@@ -1,3 +1,3 @@
-#include "detectiontool.h"
+#include "main.h"
 
 int scan_sys_call_table(void);
