+++ src/detectinodes.c
